# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "myExtBot API",
    "version": "0.1.0",
    "description": "REST API for myExtBot — an extensible multi-agent orchestration platform.\n\nAll endpoints under `/api/*` require an `x-api-key` header (set `MYEXTBOT_API_KEY` env var, default: `dev-key`).\n\nInteractive docs: `/api-docs` · Raw spec: `/api-docs/json`"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/agents/lifecycle/all": {
      "get": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Get lifecycle history for all agents",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Maximum number of history entries to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Lifecycle history records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}/status": {
      "get": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Get current lifecycle status of an agent",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Agent ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Current status record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Transition an agent to a new lifecycle status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Agent ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "initializing",
                      "active",
                      "busy",
                      "sleeping",
                      "retired"
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transition successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "record": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status or transition error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}/lifecycle": {
      "get": {
        "tags": [
          "Lifecycle"
        ],
        "summary": "Get lifecycle history for a specific agent",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Agent ID"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Maximum number of history entries to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Lifecycle history for the agent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lineage": {
      "get": {
        "tags": [
          "Lineage"
        ],
        "summary": "Get the execution lineage graph",
        "description": "Returns the full lineage graph. Use the `format` parameter to get Mermaid or DOT output.",
        "parameters": [
          {
            "in": "query",
            "name": "startTime",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter start time (ISO 8601)"
          },
          {
            "in": "query",
            "name": "endTime",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter end time (ISO 8601)"
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "mermaid",
                "dot"
              ],
              "default": "json"
            },
            "description": "Output format"
          }
        ],
        "responses": {
          "200": {
            "description": "Lineage graph in the requested format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/lineage/mermaid": {
      "get": {
        "tags": [
          "Lineage"
        ],
        "summary": "Get lineage graph as Mermaid flowchart text",
        "parameters": [
          {
            "in": "query",
            "name": "startTime",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endTime",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mermaid flowchart text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agents/route": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Route a query to the most relevant agents",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Natural language query to route"
          },
          {
            "in": "query",
            "name": "topN",
            "schema": {
              "type": "integer",
              "default": 3
            },
            "description": "Number of top suggestions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of agent routing suggestions (sorted by score)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "agentId": {
                        "type": "string"
                      },
                      "score": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}